apex/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
cryptography>=41.0.0
httpx>=0.25.0
pandas>=2.0.0
pandas_ta>=0.3.14b
yfinance>=0.2.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
alpaca-py>=0.21.0
